refactor: simplify tabular pages api
