+++ packages/application-components/src/components/internals/messages.ts
@@ -2,8 +2,12 @@ import { defineMessages } from 'react-intl';
 
 const messages = defineMessages({
   back: {
-    id: 'Components.Page.TopBar.Back',
-    defaultMessage: 'Back to list view',
+    id: 'Components.ModalPage.TopBar.Back',
+    defaultMessage: 'Go Back',
+  },
+  close: {
+    id: 'Components.ModalPage.TopBar.Close',
+    defaultMessage: 'Close Modal Page',
   },
 });
