--- packages/application-components/src/components/modal-pages/internals/messages.ts
@@ -1,14 +1,0 @@
-import { defineMessages } from 'react-intl';
-
-const messages = defineMessages({
-  back: {
-    id: 'Components.ModalPage.TopBar.Back',
-    defaultMessage: 'Go Back',
-  },
-  close: {
-    id: 'Components.ModalPage.TopBar.Close',
-    defaultMessage: 'Close Modal Page',
-  },
-});
-
-export default messages;
