+++ website-components-playground/src/components/playground-controller.js
@@ -19,8 +19,7 @@ const PreviewContainer = styled.div`
   width: 100%;
   overflow: hidden;
   height: ${(props) => props.height};
-  background-color: ${(props) =>
-    props.background || customProperties.colorSurface};
+  background-color: ${customProperties.colorSurface};
   border-radius: ${customProperties.borderRadius4}
     ${customProperties.borderRadius4} 0 0;
   border-bottom: 1px solid ${customProperties.colorNeutral90};
@@ -41,7 +40,7 @@ const PlaygroundController = (props) => {
           {({ form, values }) => (
             <PlaygroundContainer>
               <Spacings.Stack scale="s">
-                <PreviewContainer height="400px" background={props.background}>
+                <PreviewContainer height="400px">
                   {props.children({ values })}
                 </PreviewContainer>
                 <Spacings.Inline alignItems="center">
