+++ website-components-playground/src/pages/tabular-main-page.js
@@ -71,7 +71,6 @@ const TabularMainPageExample = (props) => {
               initialValue: 'default',
             },
           ]}
-          background="transparent"
         >
           {({ values }) => (
             <TabularMainPage
